feat: moving Main component to pages
